workaround the ibus window flashing issue when showing the 'w' aux hint
this patch works around the ibus window flashing when typing very fast.

this is introduced by r29, which shows hints in aux window when pressing
'w' key.

when non-'w' key pressed, ibus_engine_hide_auxiliary_text() is called,
this causes redraw, so the windows flashes when typing.

this is better fixed in the ibus core, but before that, we can work around
that in ibus-array.